+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 2.6.2, 28/10/23
 
 - better saveoptions sizing
+- add flags support to saveoptions
 
 ## 2.6.1, 12/10/23